feat: run local version when exists.
